Revert "Allow more bytes to be read from registers"
This reverts commit c0598b8f9ea289eaaef89e34eadb017c73f02298.

+++ src/SparkFunBMP384.cpp
@@ -536,7 +536,7 @@ uint8_t BMP384::readRegister(uint8_t regAddress)
     return data;
 }
 
-void BMP384::readRegisters(uint8_t regAddress, void* dataBuffer, uint16_t numBytes)
+void BMP384::readRegisters(uint8_t regAddress, void* dataBuffer, uint8_t numBytes)
 {
     switch(communicationInterface)
     {
